stdatomic: add long versions that are the same width as int
initially atomic_long, and atomic_ulong along with several other
atomic types such as atomic_bool, and atomic_intptr_t were not
added because it appears _Generic in MSVC will complain if there
are clauses that resolve to duplicate desugared types.

however, while this may be the case, it seems that long is
distinct from int even though it shares the same type width,
hence we can safely add long and ulong versions of the apis.