stdatomic: add support for non standard atomic_ptr type
while ptrdiff_t and size_t decay to their associated integer type,
this is not the case for pointers. we add an atomic_ptr type:

  typedef void * _Atomic atomic_ptr;

unfortunately this is not a complete solution because while we now
have an atomic_ptr works for void*, we still cannot support
pointers to arbitrary types unless explicitly cast to atomic_ptr.

e.g.

  typedef struct { int x; } Custom;

  void * _Atomic s;
  Custom * _Atomic u;

  void f1(void *t) { atomic_compare_exchange_strong(&s,&t,0); }
  void f2(Custom *v) { atomic_compare_exchange_strong(&u,&v,0); }

results in this error in MSVC:

  error C7702: no compatible type for 'Custom * volatile *'
  in _Generic association list

while it works in other C11 compilers supporting <stdatomic.h>.
it still seems to be worthwhile adding atomic_ptr because it is
an appropriate cast for volatile pointer types.